requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

bitsandbytes>=0.40.0
peft>=0.4.0
trl>=0.5.0

unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

numpy>=1.21.0
pillow>=8.0.0
matplotlib>=3.3.0
tqdm>=4.60.0
pyyaml>=6.0.0